refactor(rc-1): hoist closure-free helpers to lib/runtime.ts and scaffold lib/tools/
First incremental slice of the Phase 2 RC-1 split. Extracts module-level and closure-free helpers that were duplicated or trapped inside OpenAIOAuthPlugin and establishes the lib/tools/ directory + _shared.ts barrel so follow-up PRs can move each of the 18 inline codex-* tools into its own file without re-extracting the same scaffolding.

Moved to lib/runtime.ts: matchesWorkspaceIdentity, upsertFlaggedAccountRecord, normalizeToolOutputFormat, renderJsonOutput, createRetryBudgetUsage, serializeSelectionExplainability, formatRoutingValue, formatExplainabilitySummary, and the SelectionSnapshot / SerializedSelectionExplainability / RoutingVisibilitySnapshot / RuntimeMetrics / ToolOutputFormat types.

toolOutputFormatSchema and toolSensitiveJsonSchema stay in index.ts because their inferred Zod return type cannot be named without leaking @opencode-ai/plugin/node_modules/zod into the emitted declarations (TS2742). They will be threaded through ToolContext in the per-tool extraction PR.

index.ts: 6002 -> 5876 lines (-126). No public API changes. 2088/2088 tests pass unchanged.

Audit refs: docs/audits/07-refactoring-plan.md#rc-1; docs/audits/14-top20.md item #20; docs/audits/13-phased-roadmap.md section 2.1.
